qemu: remove external disk after programming internal
When both internal and external disks are defined, the flasher will
program the internal disk and then create a new EFI boot entry for the
internal disk and configure to boot from it.

However, modifying the EFI boot list is not working for QEMU, so we
still need to disconnect the external disk after programming in order to
boot from the internal.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>