qemu: flash: don't remove external disk after flashing
Secure boot tests reflash the internal storage after tampering with
verified boot files, and so we need to have the flasher media connected
to the guest as the external drive even after the internal drive is
flashed for the first time.

This snippet was intended to be used for testing the migrator, which
essentially boots and flashes to a single disk, but a race condition in
unlocking USB attached storage devices is currently blocking the
migrator from being used.

Disable removal of the external disk after the first flash.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>